Update: Hit Position
Former-commit-id: a77eb47e06a8bb075df3cd8c50f855c857cdae22 [formerly 5d13699c48897a57b10535dc1ee0eaf16f55bb1e]
Former-commit-id: a983cb60b3bf483ae895708889a506c28fc4d7d8

+++ Quaver/src/GameState/Gameplay/PlayScreen/Playfield.cs
@@ -110,7 +110,6 @@ namespace Quaver.GameState.Gameplay.PlayScreen
                 {
 
                     SizeX = GameplayReferences.PlayfieldObjectSize,
-                    
                     Position = new Vector2(GameplayReferences.ReceptorXPosition[i], GameplayReferences.ReceptorYOffset),
                     Alignment = Alignment.TopLeft,
                     Parent = Boundary
@@ -121,11 +120,11 @@ namespace Quaver.GameState.Gameplay.PlayScreen
                 {
                     case GameModes.Keys4:
                         Receptors[i].Image = GameBase.LoadedSkin.NoteReceptors4K[i];
-                        Receptors[i].SizeY = GameplayReferences.PlayfieldObjectSize * (float) GameBase.LoadedSkin.NoteReceptors4K[i].Height / GameBase.LoadedSkin.NoteReceptors4K[i].Width;
+                        Receptors[i].SizeY = GameplayReferences.PlayfieldObjectSize * GameBase.LoadedSkin.NoteReceptors4K[i].Height / GameBase.LoadedSkin.NoteReceptors4K[i].Width;
                         break;
                     case GameModes.Keys7:
                         Receptors[i].Image = GameBase.LoadedSkin.NoteReceptors7K[i];
-                        Receptors[i].SizeY = GameplayReferences.PlayfieldObjectSize * (float) GameBase.LoadedSkin.NoteReceptors7K[i].Height / GameBase.LoadedSkin.NoteReceptors7K[i].Width;
+                        Receptors[i].SizeY = GameplayReferences.PlayfieldObjectSize * GameBase.LoadedSkin.NoteReceptors7K[i].Height / GameBase.LoadedSkin.NoteReceptors7K[i].Width;
                         break;
                 }
             }
